[security] API routes : refactor server logs and data validation

+++ src/pages/api/goals/[id].ts
@@ -4,14 +4,15 @@ import { db, Goals, eq } from 'astro:db';
 export const DELETE: APIRoute = async (ctx) => {
     try {
         const id = Number(ctx.params.id);
-        if (!id) {
-            return new Response('Error : id is null',{ status: 400 });
+        if (!id || id < 1 || !Number.isInteger(id)) {
+            return new Response('Error : invalid id',{ status: 400 });
         }
         await db.delete(Goals).where(eq(Goals.id,id));
         return new Response(null, { status: 204 });
     }
     catch (error) {
-        return new Response(`Error: ${error.message}`,{status: 500});
+        console.error(error);
+        return new Response('An error occurred',{ status: 500 });
     }
 };
 
@@ -19,8 +20,8 @@ export const PATCH: APIRoute = async (ctx) => {
     try {
         const id = Number(ctx.params.id);
         const { updatedCompleted } = await ctx.request.json();
-        if (!id) {
-            return new Response('goal id not foundl',{ status: 400 });
+        if (!id || id < 1 || !Number.isInteger(id)) {
+            return new Response('Error : invalid id',{ status: 400 });
         }
         if (typeof updatedCompleted !== 'string') {
             return new Response('invalid completed value',{ status: 400 });
